project/requirements.txt
 Django>=2.0,<3.0 
 django-crispy-forms==1.9.0 
 djangorestframework==3.9.0 
 djangorestframework_simplejwt==4.1.4 
 requests
 django-cors-headers==3.4.0
